+++ src/stories/BeamContainer.tsx
@@ -4,20 +4,13 @@ import React, {
   ReactNode,
   isValidElement,
 } from "react";
-import { Note, NoteProps } from "./Note";
+import { Note } from "./Note";
+import { noteFlexValue } from "../helpers/helpers";
+import { NoteProps } from "../helpers/types";
 
 interface BeamContainerProps {
   children?: ReactElement<typeof Note> | ReactElement<typeof Note>[];
 }
-
-//! noteFlexValue is used in Note.tsx - need to put it in a helper file
-const noteFlexValue: Record<NoteProps["noteValue"], number> = {
-  whole: 16,
-  half: 8,
-  quarter: 4,
-  eighth: 2,
-  "16th": 1,
-};
 
 //! I don't fully understand wha't happening here, need to review
 const isNoteElement = (child: ReactNode): child is ReactElement<NoteProps> => {
