+++ src/stories/Note.tsx
@@ -1,144 +1,15 @@
 import React from "react";
 import "./Note.css";
 import "../global.css";
-import { NoteGlyphs, accidentalGlyphs, noteGlyphs } from "../helpers/glyphs";
-
-export type PitchPosition =
-  | "line-above-1"
-  | "space-above-1"
-  | "line-5"
-  | "space-4"
-  | "line-4"
-  | "space-3"
-  | "line-3"
-  | "space-2"
-  | "line-2"
-  | "space-1"
-  | "line-1"
-  | "space-below-1"
-  | "line-below-1";
-
-interface BeamPositionsType {
-  [key: string]: { [key: string]: number };
-}
-
-const BeamPositions: BeamPositionsType = {
-  upStem: {
-    "line-above-1": 12,
-    "space-above-1": 16,
-    "line-5": 20,
-    "space-4": 24,
-    "line-4": 28,
-    "space-3": 32,
-    "line-3": 36,
-    "space-2": 40,
-    "line-2": 44,
-    "space-1": 48,
-    "line-1": 52,
-    "space-below-1": 56,
-    "line-below-1": 60,
-  },
-  downStem: {
-    "line-above-1": 64,
-    "space-above-1": 68,
-    "line-5": 72,
-    "space-4": 76,
-    "line-4": 80,
-    "space-3": 84,
-    "line-3": 88,
-    "space-2": 92,
-    "line-2": 96,
-    "space-1": 100,
-    "line-1": 104,
-    "space-below-1": 108,
-    "line-below-1": 112,
-  },
-};
-
-const StemPositions = {
-  "line-above-1": 40,
-  "space-above-1": 44,
-  "line-5": 48,
-  "space-4": 52,
-  "line-4": 56,
-  "space-3": 60,
-  "line-3": 64,
-  "space-2": 68,
-  "line-2": 72,
-  "space-1": 76,
-  "line-1": 80,
-  "space-below-1": 84,
-  "line-below-1": 88,
-};
-
-interface BaseNoteProps {
-  noteValue: "whole" | "half" | "quarter" | "eighth" | "16th";
-  // | "32nd"
-  // | "64th"
-  // | "128th"
-  // | "256th"
-  // | "512th"
-  // | "1024th";
-  dotted?: 1;
-}
-
-interface Beam {
-  amount: 1 | 2;
-  status: "start" | "continue" | "end";
-  hook?: "forward" | "backward";
-  nextNotePostion?: PitchPosition;
-}
-
-interface RestProps extends BaseNoteProps {
-  rest: true;
-  pitch?: never;
-  position?: PitchPosition;
-  stem?: never;
-}
-
-interface NoteValueProps extends BaseNoteProps {
-  pitch?: {
-    step?: "A" | "B" | "C" | "D" | "E" | "F" | "G";
-    alter?: "sharp" | "flat" | "natural" | "doubleSharp" | "doubleFlat";
-    octave?: 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
-  };
-  position: PitchPosition;
-  stem?: "upStem" | "downStem" | "noStem";
-  rest?: false;
-  beam?: Beam;
-}
-
-export type NoteProps = RestProps | NoteValueProps;
-
-const noteTranslations: Record<NoteProps["noteValue"], keyof NoteGlyphs> = {
-  whole: "wholeNote",
-  half: "halfNote",
-  quarter: "quarterNote",
-  eighth: "eighthNote",
-  "16th": "sixteenthNote",
-};
-
-const noteFlexValue: Record<NoteProps["noteValue"], number> = {
-  whole: 16,
-  half: 8,
-  quarter: 4,
-  eighth: 2,
-  "16th": 1,
-};
-
-const getDefaultStem = (position: PitchPosition): "upStem" | "downStem" => {
-  const downStemPositions: PitchPosition[] = [
-    "line-above-1",
-    "space-above-1",
-    "line-5",
-    "space-4",
-    "line-4",
-    "space-3",
-    "line-3",
-  ];
-
-  return downStemPositions.includes(position) ? "downStem" : "upStem";
-};
+import { accidentalGlyphs, noteGlyphs } from "../helpers/glyphs";
+import {
+  BeamPositions,
+  getDefaultStem,
+  noteFlexValue,
+  noteTranslations,
+  StemPositions,
+} from "../helpers/helpers";
+import { Beam, NoteProps } from "../helpers/types";
 
 export const Note = (props: NoteProps) => {
   const { noteValue, rest } = props;
